Factor out RangeValidator out of BetweenValidators
